refactor(careerExp): migrate to tailwind
