+++ src/screens/CV/components/TechnicalSkills/TechnicalSkills.tsx
@@ -18,7 +18,7 @@ export const TechnicalSkills = ({ technical_skills }: TechnicalSkillsProps) => {
           return (
             <div key={id} className="flex flex-col flex-wrap align-baseline">
               <h3 className="font-bold">{name}:</h3>
-              <ul className="flex-1 flex flex-row space-x-1.5">
+              <ul className="flex-1 flex flex-row space-x-1.5 flex-wrap">
                 {technologies!.map((tech, index) => {
                   const { id, name } = tech;
                   const isNotLast = index !== technologies.length - 1;
