--- src/screens/CV/components/shared/Project/styles.ts
@@ -1,51 +1,0 @@
-import { styled, css } from '@screens/CV/styled';
-
-export const Wrapper = styled.article`
-  ${({ theme }) => css`
-    &:not(:last-child) {
-      margin-bottom: ${theme.pxToRem(16)};
-    }
-
-    /* Tags generated by react-markdown */
-    > ul li {
-      display: block;
-      padding-left: 16px;
-      margin-bottom: 4px;
-      &::before {
-        content: '• ';
-      }
-    }
-  `}
-`;
-
-export const Title = styled.h3`
-  ${({ theme }) => css`
-    font-size: ${theme.pxToRem(16)};
-    font-weight: bold;
-
-    @media (min-width: ${theme.sizes.tablet}) {
-      font-size: ${theme.pxToRem(20)};
-    }
-  `}
-`;
-
-export const Info = styled.div`
-  ${({ theme }) => css`
-    margin-bottom: ${theme.pxToRem(0.8)};
-
-    @media (min-width: ${theme.sizes.tablet}) {
-      justify-content: flex-start;
-      flex-direction: row;
-    }
-  `}
-`;
-
-export const Subtitle = styled.h4`
-  flex: 1;
-  font-weight: 400;
-`;
-
-export const Period = styled.p`
-  font-size: ${({ theme }) => theme.pxToRem(14)};
-  opacity: 0.6;
-`;
